requirements.txt
jiwer
st-supabase-connection
spellchecker